LayoutTests/imported/w3c/web-platform-tests/content-security-policy/generic/304-response-should-update-csp.sub-expected.txt

CONSOLE MESSAGE: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

PASS Test that the first frame uses nonce abc
NOTRUN Test that the first frame does not use nonce def
FAIL Test that the second frame uses nonce def assert_unreached: Unexpected message received Reached unreachable code
FAIL Test that the second frame does not use nonce abc assert_unreached: Unexpected message received Reached unreachable code
